Removed padding from dashboard, moved PendingLeasesTable into tables folder
