+++ front-end/src/components/tables/PendingLeasesTable.jsx
